+++ src/main/java/com/huadao/community/controller/QuestionController.java
@@ -19,7 +19,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable Integer id, Model model) {
+    public String question(@PathVariable Long id, Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
         model.addAttribute("question", questionDTO);
